+++ cmd/gapic-showcase/repeat-data-body.go
@@ -56,6 +56,8 @@ var repeatDataBodyInputInfoPChildPBool bool
 
 var RepeatDataBodyInputInfoPChildPContinent string
 
+var repeatDataBodyInputIntendedBindingUri string
+
 var repeatDataBodyInputPInt32 int32
 
 var repeatDataBodyInputPInt64 int64
@@ -189,6 +191,8 @@ func init() {
 
 	RepeatDataBodyCmd.Flags().BoolVar(&RepeatDataBodyInput.ServerVerify, "server_verify", false, "If true, the server will verify that the received...")
 
+	RepeatDataBodyCmd.Flags().StringVar(&repeatDataBodyInputIntendedBindingUri, "intended_binding_uri", "", "The URI template this request is expected to be...")
+
 	RepeatDataBodyCmd.Flags().Int32Var(&RepeatDataBodyInput.FInt32, "f_int32", 0, "Some top level fields, to test that these are...")
 
 	RepeatDataBodyCmd.Flags().Int64Var(&RepeatDataBodyInput.FInt64, "f_int64", 0, "")
@@ -294,6 +298,10 @@ var RepeatDataBodyCmd = &cobra.Command{
 
 			if cmd.Flags().Changed("info.p_child.p_bool") {
 				RepeatDataBodyInput.Info.PChild.PBool = &repeatDataBodyInputInfoPChildPBool
+			}
+
+			if cmd.Flags().Changed("intended_binding_uri") {
+				RepeatDataBodyInput.IntendedBindingUri = &repeatDataBodyInputIntendedBindingUri
 			}
 
 			if cmd.Flags().Changed("p_int32") {