+++ server/genrest/testing.go
@@ -19,6 +19,7 @@ package genrest
 
 import (
 	"bytes"
+	"context"
 	genprotopb "github.com/googleapis/gapic-showcase/server/genproto"
 	"github.com/googleapis/gapic-showcase/util/genrest/resttools"
 	gmux "github.com/gorilla/mux"
@@ -89,7 +90,8 @@ func (backend *RESTBackend) HandleCreateSession(w http.ResponseWriter, r *http.R
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.CreateSession(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/sessions")
+	response, err := backend.TestingServer.CreateSession(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -150,7 +152,8 @@ func (backend *RESTBackend) HandleGetSession(w http.ResponseWriter, r *http.Requ
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.GetSession(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{name=sessions/*}")
+	response, err := backend.TestingServer.GetSession(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -206,7 +209,8 @@ func (backend *RESTBackend) HandleListSessions(w http.ResponseWriter, r *http.Re
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.ListSessions(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/sessions")
+	response, err := backend.TestingServer.ListSessions(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -267,7 +271,8 @@ func (backend *RESTBackend) HandleDeleteSession(w http.ResponseWriter, r *http.R
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.DeleteSession(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{name=sessions/*}")
+	response, err := backend.TestingServer.DeleteSession(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -328,7 +333,8 @@ func (backend *RESTBackend) HandleReportSession(w http.ResponseWriter, r *http.R
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.ReportSession(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{name=sessions/*}:report")
+	response, err := backend.TestingServer.ReportSession(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -389,7 +395,8 @@ func (backend *RESTBackend) HandleListTests(w http.ResponseWriter, r *http.Reque
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.ListTests(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{parent=sessions/*}/tests")
+	response, err := backend.TestingServer.ListTests(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -450,7 +457,8 @@ func (backend *RESTBackend) HandleDeleteTest(w http.ResponseWriter, r *http.Requ
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.DeleteTest(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{name=sessions/*/tests/*}")
+	response, err := backend.TestingServer.DeleteTest(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return
@@ -511,7 +519,8 @@ func (backend *RESTBackend) HandleVerifyTest(w http.ResponseWriter, r *http.Requ
 	requestJSON, _ := marshaler.Marshal(request)
 	backend.StdLog.Printf("  request: %s", requestJSON)
 
-	response, err := backend.TestingServer.VerifyTest(r.Context(), request)
+	ctx := context.WithValue(r.Context(), resttools.BindingURIKey, "/v1beta1/{name=sessions/*/tests/*}:check")
+	response, err := backend.TestingServer.VerifyTest(ctx, request)
 	if err != nil {
 		backend.ReportGRPCError(w, err)
 		return